[opt] recode updater to ts
